requirements.txt
rich==13.7.0
prompt_toolkit==3.0.43
pyyaml==6.0.1
python-dotenv==1.0.0
watchdog==3.0.0
colorama==0.4.6
tqdm==4.66.1
google-generativeai==0.3.2
# New dependencies
psutil==5.9.8
pyttsx3==2.90
SpeechRecognition==3.10.0
pyaudio==0.2.14
pillow==10.2.0
pypdf2==3.0.1
requests==2.31.0
beautifulsoup4==4.12.2 